Fix:메인 서비스 파일 병합
